Flow manager bug fixed:
When the flow was deleted from the flow manager, it would delete all of
the flows with that title. We changed it so that it deletes the only
flow being removed. To test this, add multiple flows with the same
title, and observe that only one of them is removed.