Pop the route transition params off the stack later in the lifecycle.
When a user visits a card from an existing page (i.e. flow manager) the
app caches that pages model.  When the user closes the card the app will
use that cache to avoid having to reload the previous page.  The
redirect params entry was being popped off of the stack too early in the
transition, and later methods wouldn't know to look for the cached
model.