Refactor out modals for every user in user search
User details were being rendered in separate hidden modals,
which was potentially creating a lot of DOM elements most of which would
never be shown. We moved to using the application-level modal and
re-using it for each user that was clicked on. This makes the user
search and detail viewing more snappy.