Improve DOI lookup
Previously, when given a DOI, the Paper model was calling find
with an ID, and failing with a non-id lookup. This was problematic. One,
if the journal prefix happened to be the same as a paper, this would
have created bogus paper entries. Additionally, two lookups were
required to get a Paper record.

The current approach is to:

1. Check numericality of the id param

2. If it's an integer, perform the lookup based upon an id value.
Otherwise, continue to lookup the Paper using the doi.