Switch back to using selenium. Deemed lesser of two evils (see below)
We’re using a sinatra application that mocks out event stream interaction. To accomplish this, we setup the thin application server (event machine). This works fine for selenium tests, but when combined with capybara-webkit, the tests slow to a halt (taking roughly ~1 minute per test).

So, our options given the current state of ember-cli-rails:

1) separate tests into two distinct runs. One using thin, and one using capybara default (via rspec tags)
2) Move capybara webkit tests back to selenium (incur test speed decrease)

Option 2 was chosen because it requires less orchestration and the problem itself goes away if ember-cli-rails fixes issue #71 (https://github.com/rwz/ember-cli-rails/issues/71). Or, we decide to switch *just* ember-cli and separate the client from the api. At the moment the work required to do that is out of scope (switching to token based authentication, changing deploy process).