Upgrade fog to remove warning output (see below)
Ideally we could just use `fog-aws` since that's the only module
we're using, but the library currently seems to be in such a state
that that's not immediately possible. However, it is a goal
of the project. After speaking with a rackspace employee, he recommended
still including `fog`.

Follow https://github.com/fog/fog/issues/3279 for more info

In the maintainers words, "It's going to get worse before it gets
better".