Fix Add Authors feature spec
The specs inside the "with an existing author" context all rely on an
existing author. The author is created with `let!` as soon as we enter
the context. However, in the `before` block outside of the context we
navigate to the edit page.

The browser is now looking at the edit page and *then* the author is
created. This change in the data on the back end doesn't propagate up to
the browser. When the authors card is opened, the expected author isn't
there, and the specs fail.

In this change we move the act of navigating from the dashboard to the
edit page to inside each scenario. This way we load the edit page after
all its data has been set up on the back end.