Force a reload on the task for every task.
Currently if Task A relies on data that is modifiable by Task B,
Task A won't necessarily get the freshest data since `store.findTask` is
grabbing the task from the local cache. This seems to be relevant especially for
collections. In this case, completing "Register Decision" creates a decision on the
backend, which "Invite Reviewers" relies on. When the user exists Register Decision
and opens Invite Reviewers, the latest decision is not in the data store.