Give card-thumbnail a comment looks relationship
* So we don’t have to pass all the comment looks around, instead
scope to just that task (or in this case, card-thumbnail).

* Since card-thumbnail is an ember only model, in order to make this
relationship work, I had to implement a custom normalize method for
the comment-look serializer. Ideally the card-thumbnail wouldn’t be 
necessary and we could just use a sparse view of the task - having the
data upgraded from the server when the user requests more data.