Eventstream PaperRoles to the Paper channel
If you only send the paper role to the user, no one else is notified
that the new user has a new role.

For example, after an Editor accepts their invitation, the PaperRole is
created and they gain access to the Paper. The Admin who invited them is
needs to also be notified that the Editor accepted.