Clean up and clarify .gitignore; better .env
I have removed .DS_Store; please add it to your `~/.gitignore_global`.

According to the best practice suggestions outlined in the
[dotenv](https://github.com/bkeepers/dotenv) docs, `.env` itself has
been populated with global development defaults that are committed and
shared across machines, and the setup script now puts a machine-unique
Rails secret token base in `.env.local`, which is not committed.

`.env.local` is the only ignored file, since `.env` is effectively for
development and `.env.test` was excluded before anyway.