Raise error when in an unrecoverable situation
If the User record itself is invalid after login, there isn’t anything
the user (human) can do to correct it. They certainly can’t do much if
we actually logged them in.

Raise an error to the user and to Bugsnag. This should give us the info
we need to debug our issue.