Avoid pointlessly serializing Paper for Heartbeats
Not sure why we were doing this. The Ember Heartbeat service never uses
the result it gets back from the heartbeat request. It just drops it on
the ground. It uses the RESTless service and isn't pushing anything into
the Store. No one should even need the data from the heartbeat, as that
data would be eventstreamed to them anyway.

This is an issue since the PaperSerializer serializes everything under
the sun. This should take this endpoint from being one of the most
memory/time intensive to just a blip on the radar.