Resolve intermittent invite reviewer spec failure, e.g.:
   rspec engines/tahi_standard_tasks/spec/features/invite_reviewer_task_spec.rb 

MORE INFO:

Make sure we don't attempt to invite the reviewer until after the current task has been updated.

Otherwise, we create a race condition where `inviteReviewer` tries to create an invitation while the update request for the task is in progress. 

If the invitation is created but not in the right state (basically we hit line 12 but haven't yet hit line 13 in invitations_controller.rb) then the update to the task will include a serialized "pending" invitation.

At this time if ember-data is waiting for the creation of the invitation to resolve (e.g. the "then" block passed in at line 44 in paper-reviewer.coffee) then ember-data will resolve it when it sees the pending invitation in the JSON response from updating the task. 

This seems like a bug in ember-data as it shouldn't resolve it until it sees the invitation in the response from the POST that creates an invitation. Or, it may be a feature in how ember-data tries to be smart about loading associated models on the client side.

Another possible bug in ember-data is that when the POST that creates the invitation does respond ember pretty much ignores it since it has already resolved the promise prematurely. It seems like it should update the data-store and propagate out any changes to computed properties, etc.

OTHER POSSIBLE FIXES

This may not be the best solution to this problem. Here are a few other possibilities:

* Removing the "has_many :invitations" line in the paper_review_task_serializer.rb also fixes this issue because ember will no resolve the promise early

* Wrapping this action in another layer of transactions may fix the issue by not allowing a partially "created" invitation to leak out in another request

* Other ideas....?