An invitation's body is now stored on the invitation itself rather than the task that it is associated with.
Fixes a bug where a PUT to task would set the invitation body, then a separate POST to invitations would create the invitation and delay it for sending, and then when it was sent it would try to pull the body off the task. However, the task may have been changed causing the body to be updated so that the invitee got the wrong invitation body.

* No longer need to save the task when sending an invitation. Everything needed to create/send an invitation should come thru the POST /api/invitations end-point

[#86260910]