Attach Pusher's socket id to RESTless calls
The server expects all incoming requests to have a socket id. This is
used by the Broadcaster to exclude the original requester from having
events they already know about pushed to them. It is important to
exclude the originator to prevent UI flashing as the request and the
websocket data are both pushed into the store.