Updating auto-suggest component to fire inputChanged on keyup rather than on text changed.
The reason is that we only want to fire this when the user types something in. Here's what was happening:

* user clicks on "John Doe [john@example.com]" search result
* itemSelected is fired
* farther up on the consuming code, itemSelected is used to set the text value to "John Doe [john@example.com]"
* which causes resultTextChanged to get fired
* which causes the consuming code farther up the chain to receive inputChanged

We don't want the last two bullet points above occurring when the user clicks on or selects the a search result. By listening to keyUp we can limit this chain of events to only when the user is typing.