[#100589444] right/ctrl click to open card in new tab
Right now, the only way to open a card in another
tab is to copy and paste the card URL into a new
tab you open manually. You should be able to right
click or ctrl click on a card, and have an option
to open that card in a new tab.

This is done by making the card-preview component
an anchor element and generating the href.

I didn't want to to hard wire the href so I ended
up introducing a hack way of getting to the Router
from a Component. I tried injecting the router into
the card-preview component from an initializer
but this ended up causing problems in the tests.

Also slipped in a fix for a deprecation warning
on the Discussion title input.