Fix dragging of card-previews
A previous PR changed card-preview components into anchor elements to
allow control/right clicking on them to open in a new tab. It turns out
jQueryUI sortable freaks out when the draggable element is an anchor.

The fix was to change the containing element for card-preview back to a
div and change a child element into an anchor. The style and markup
changes are minimal.