Remove select-box background cover, open with spacebar
Add opening with spacebar when focused to mimic
native select elements.

Removed the background cover because css z-index.
When a select-box was in an overlay (or any other
element with a z-index set) the cover would either
sit on top of everything or would be below everything.
In place of this a click observer is put on the
document.