Fixing broken specs: collaborations_controller_spec.rb and add_collaborators_spec.rb
This branch removed the route because there was no controller backing it:

     resources :paper_roles, only: [:show]

This route masked an issue where CollaborationsController is responding with a PaperRole object. ActionPack by default will try to use polymorphic routing to set the Location header, e.g.: url_for(resource) when using the respond_with helpers.

So, we were sending the client back a Location that didn't actually resolve to a route that we supported (we had a definition, but not a controller to respond to it).

This sets the location specifically to nil since our Collaboration resource doesn't support a "show".