Remove unnecessary callback
This appears to be here in order to work around a broken association
between Tasks and Invitations. Even when Ember Data sees an Invitation
with a specific task_id, it won't hook up the relationships correctly.

We fixed this in a slightly more general way, for both the original
user, as well as anyone listening to the Paper channel. Now when an
Invitation is created/updated, we include the entire Task in the
serialized Invitation. This is itself a bit of a hack. Maybe a future
version of Ember Data will handle this case correctly.