Avoid notifying non-existent Invitees
When Invitations are created, they are created in a pending state. The
client sends an email address only, and the Invitation itself is
responsible for linking up an existing User. This means that Invitations
are created _without_ an `invitee` set. This is fine, since the
controller immediately calls `Invitation#invite!`, which does the state
machine transition and links up existing users.

So Invitations are never _created_ with an `invitee`. There is no need
to attempt to EventStream anything to this undefined user. If they are
an existing User and logged in, they will be notified after the state
machine transitions to `invited` and fires the "invitation:updated"
event.