WIP / Exploration.
This removes the display-line-break.js and line-break-to-tag.js as we can rely on CSS to preserve whitespace handling.

In many cases we can use HTMLBars triple brace {{{...}}} to render content as-is. Modern browsers support not executing <script> tags. Reference: https://developer.mozilla.org/en-US/docs/Web/API/Element/innerHTML . BUT this is breakable with content like:

     <iframe src=""srcdoc="<script>alert('hi');</script>"></iframe>

Currently, some user input-able content is stored as plain text on the backend. Other times, it's converted to simple HTML on the front-end and saved in the backend.

This means a bunch of the {{{...}}} in this PR may not be able to be trusted. Need to circle back with team and find a uniform way of handling user inputtable content.