Match standard Rails as_json method signature
When the letter template is being rendered, it is being passed around a
lot of internal Rails rendering methods. Those internals expect the
standard `as_json` signature. They are attempting to pass an options
hash to `LetterTemplate#as_json`, but instead explode with an
ArgumentError. LetterTemplate doesn't need to necessarily do anything
with those options, it just needs to not explode so spectacularly
whenever a Task is updated.