Update Dashboard after removing a collaborator
This is a fairly infrequent event, but could be quite confusing to
experience as a user.

If a collaborator is removed from the Paper while they are on their
Dashboard, remove the Paper. We only subscribe to the User channel while
on the Dashboard, so this User was missing out on the event that removed
them.