Add attribute to record full record namespace
Even though EventStore has a polymorphic relationship, rails only stores
the base class name to the related record, not the full namespace.

This is especially painful for a subclassed Task where we might want
to search across all EventStore records for a particular Task type.
So, we added an explicit field to store this information.