Remove TODO comments
The suggested AR relationship reversal is not a viable solution.  In 
order for an AuthorsTask to be marked complete, it must be valid.
Validity for an AuthorsTask is currently defined by business
requirements to be that all authors from that task are valid.

Therefore, as it stands now there needs to be some connection between
Authors and an AuthorTask.  This unfortunately couples the two models
too tightly, but deemed out of scope for this current work to fix
at this moment.

We are hopeful that future work on how validations do -- and should --
work will make this code irrelevant.