Updating NestedQuestionAnswer to handle update and destroy actions.
* because ember and rails have different philosophies on polymorphic associations we have to introduce picking logic for finding the correct "owner_type" for an answer. *sigh*. 

     E.g. ember submits "Funder" but we have TahiStandardTasks::Funder

I am not sure how pluggable architecture was ever going to work for things not embedded directly into core.

In an ideal world it seems like the ember-side models would be able to be looked up via "tahi-standard-tasks/funder" so it would resubmit that as a value to the server. *shrug*