Correctly subscribe to Discussion Topics
DiscussionTopicsPolicy required that a paper be passed in during
initialization. This was *typically* the case, but there were two
situations where this fell down:

1. When the policy needed to check `index?`
2. When the EventStream needed to check `show?`

The current policy framework does not well support situations like this
and might need reevaluated. However, that work falls outside the scope
of this bug fix.