Pusher pushes only ID and type.
We currently have to serialize every model *twice* when we PUT, once to
return and once to send down pusher.

This is not only slow, but also a major security issue, as we end up
pushing sensitive information to to broad a user-base.

This will require setting up endpoints for many models that don't
currently have `show` routes.