Add the concept of "previous decision" to ReviewerReportTask.
This is so it can know which list of decisions apply to it. Without this the previous decisions were being loaded off of the paper which is not correct because that will load decisions for other ReviewerReportTask(s) that exist on the paper (but are for other reviewers to fill out).

[Pivotal Card #101433534]