Only show participation activity on workflow feed
When a Reviewer is invited to review a Paper, a ReviewerReportTask is
created for them. That Reviewer is added to that task as a Participant.
This creates a Activity record, recording this action.

It was deemed useful for the staff to add the Reviewer's name to that
Activity record. Unfortunately, it is a huge privacy hole for the anyone
else. The Author absolutely should not see who is reviewing their Paper.

To hide this from them, only publish these types of Activities to the
workflow feed, not the manuscript feed.