Turn on Capybara.wait_on_first_by_default so specs that use #first(...) will use Capybara's traditional waiting behavior.
This is to resolve an intermittent spec failure in spec/features/journal_administration_spec.rb where it cannot find a button using #first(...) in order to click on it.

Additional musings: The app should not need to use #first(...) but it uses spans and divs for representing clickable areas (as opposed to buttons or anchor tags). Because of this it can't take advantage of Capybara's #click_button, #click_link, #click_on helper methods which automatically provide the waiting behavior.