NestedQuestionAnswersController#create should ensure that we don't create duplicate answers, so it tries to find an existing answer on every POST #create.
This happened because entering a funder's role description on FinancialDisclosure would do a second POST request and create an extra record the first time you entered a value.

I do not love this, but it makes sense should two clients try to construct two answers near the same time. We always want the most recent answer.