APERTA-3495 Updating typed item storage
Items that are of a specific type (attachment, author, figure) now 
store under a single object with type properties and a name of the
type.  For example, an attachment now renders as:

{name: "attachment", type: "properties", children: [
	{name: "file", type: "text", value: "filename.ext"},
	{name: "title", type: "text", value: "file title"},
	{name: "caption", type: "text", value: "file caption"},
	{name: "status", type: "text", value: "processing"}
]

This is to make generic snapshot diffing easier.  That is, the
differ doesn't need to know about author, figure, attachment, etc,
types and can just diff text, booleans, properties, and children.