Add back in Capybara.default_max_wait_time to 10 seconds.
The default is 2 seconds which sometimes is too short for intermittently causes execution errors.

This was removed in d6fd3b3d500c0a3841775a8b8a1a44a3c386025b which works most of the time for most tests, but every now and then selenium / firefox seem to take its sweet time. 

This is the "maximum" wait time, not the minimum wait time so it shouldn't adversely affect the duration of every test run.

APERTA-5419