Updating Subscriptions to support reloading its configuration.
  * Renaming Subscriptions.clear_all_subscriptions! => unsubscribe_all for consistency

This is because the config/initializers/subscriptions.rb for each of the engines wasn't being loaded for feature specs. Rather than hard-code all of their paths or try to be clever with looking for those files on the file-system it seems simpler/cleaner to store the configuration blocks and to re-evaluate them.

APERTA-3653