Remove cookie setting as part of basic_auth check
This fixes the issue where users are basic auth prompted when
clicking the sign out link in aperta because their _tahi session
cookie is destroyed and then they are redirected to the sign in page.

Enforce authentication on ember#index page so that user is not
basic auth prompted upon redirects.