Fix broken image anchors for ingested docs
When a docx that has embedded images is uploaded from tahi, iHat will return an epub with the converted html and any embedded images will be extracted and placed into an `images` folder.  At this point, the html will contain links (<a href>) to these locally stored images.  These images will be converted to Figures so that they appear in the FiguresTask UI.

Prior to setting this html as the new paper body, the image links must be modified to point to the newly created Figure URL.

wip