Send EventStream when paper#body changes
Although the paper has a #body= setter, the body is actually stored on the corresponding VersionedText.  This change will correctly send an updated event stream notification when the paper body changes.  Otherwise, if just the #body on the paper changes (and nothing else), users would not see the latest paper body in their client.