APERTA-5690 Updating test to wait for uncompleted state
Added the ensure_not_completed method.

This method takes advantage of Capybara's default wait time to ensure
that the checkbox is in the state we want. Without expecting the state
Capybara would return the checbox right away since it is on the page.
By expecting the state in the checkbox selector we can find the checkbox
we're looking for without having to sleep.