APERTA-3611 Shows diff icon.
Still WAY slow and the positioning isn't right.

We need to change from using task.getSnapshotEtc (because ember data
isn't smart enough to be told NOT to go fetch the snapshots for *each*
task.) and instead find each snapshot via the paper.

We also need to rearrange card preview to *flow* the icons, instead of
absolutely positioning them; that way we can match the mock. Absolute
positioning is the last resort, folks!