Makes nested question seeds idempotent and powerful
We need to be able to change/add/remove nested questions,and both rake
tasks and data migrations seem sloppy. We already have to change the
seeds for nested questions, anyway, and it seems sensible to *only*
change those seeds.

This allows us to change the seeds, and use those seeds to
update/add/delete nested questions on running instances.