APERTA-5549 Removes title h2 from above manuscript
Because we extract the title from the manuscript, if we show the title
and *then* the manuscript, the title appears twice. The simplest way out
of this predicament is to show *just* the manuscript. This means that
The title may not be consistently styled; it's something we need to
think about.

The alternative would be to remove the title from the converted HTML, so
that we can *re*display it. This is complicated and dangerous and so is,
in my mind, an option to take only if this simpler way fails.