APERTA-5756 Delete state for SI files
We had some concerns about the existing delete confirmation paradigm --
namely, that the full-green overlay concealed *what* the user was
deleting, and that it required a larger amount of space that was
available in the new accordion view (the new SI file design also has
less padding, etc).

In this commit we propose a new delete confirmation, based on a mobile
pattern of slide-y buttons.

-- with AJ Hekman <aj@mutuallyhuman.com>