Change how QueryAgainstAuthorization works when querying against active record relations.
Sam (@diiq) asked a question the other day that prompted this change. He asked if we added a "limit(X)" clause to the end of a query chain, would the authorization system respect it? 

It would have, but in the wrong way so this change fixes that. 

Prior to this change the query happened against the object you were assigned to and then the object you were looking for was added as a join. So in the below call if you got access to tasks thru Paper it would construct a query against paper and join :tasks:

      user.filter_authorized(:view, Task.all) 

That worked fine, but as soon as you did:

      user.filter_authorized(:view, Task.all.limit(10))

The query wouldn't make as much sense because it'd be limiting the result of papers. You could workaround the issue by adding a where condition to ignore Task records that had a NULL id, but that seems like a poor work around and requires the caller to know more about the internals of authorization. We want the caller to know as little as possible outside of the interface they're conforming to.

So this change refactors the query so the passed in scope (e.g. Task.all.limit(10)) remained the base query and then add the necessary joins for access thru paper. This actually makes the code simpler in my opinion.

This also seems like a better choice because it is what a caller would likely intuitively think would happen and it allows us to have to perform less magic or mucking with the original query from the caller.