Adds a direct relationship from Task to Paper rather than going thru Phase.
Previously, tasks were tied to papers thru phases. This made for a more complicated query and a roundabout way of for paper to access their tasks and vise versa.

Tasks never change papers so to simplify queries, improve performance, and make the relationship more intuitive task is now stamped with its paper_id.

Another consideration is that a phase is a way to group tasks on a paper for workflow needs and it seems orthogonal to the relationship that task has with paper.

[APERTA-5890]