Updating code comment reference.
Additional note: this is because ember-data expects resources on the server to have numeric IDs by default otherwise it gets angry.

Rather than work against the grain right now we're working with it. This may change in upcoming work if this has cause other issues. Perhaps to not use ember-data for permissions on the client-side (maybe restless) and then we could get rid of a fake id.

[APERTA-5538]