Move up the creation of the authorization DB tables into a before(:all).
This so they are run before any records are attempted to be created. In some cases this was failing on CircleCI.

Remove any caching of whether or not create_db_tables was called. This is to help CircleCI not run into issues the forking or spawning processes.

This commit is somewhat of an experiment.

[APERTA-5538]