Updating TasksController to now use R&P permissions for enforce_policy
This is so we can hide the billing task for those who do not have permission to see it. This may break other tasks. If so we'll want to add new roles and permissions and make sure users are assigned appropriately rather than backtracking to the old policy implementation.


APERTA-5629