Remove TaskType class
The TaskType class was used to track "registered" task types, but it
could get out of sync in development, and was unnecessarily complicated.

Replace it with the use of two new methods on Task: all_task_types and
safe_constantize and the use of class constants for defaults, e.g.
DEFAULT_TITLE as a constant on a class.