Add decisions to every paper in the seed data.
This is so the paper in the seeds resembles what happens when a paper is created thru the UI. Also, without a decision you cannot invite reviewers.

If this seems strange keep in mind that Aperta creates decision before a decision is made. It's always one decision ahead of an actual "Register Decision" occurring.