Update Task to require_dependency on every single task in the system.
This is disgusting, but the Rails core team has decided not to pursue fixing the issue of reloading as it relates to STI. This has been an issue for over 5 years: https://github.com/rails/rails/issues/3364

This is an issue in development whenever the app reloads. This breaks R&P queries as it relies on being able to determine the descendants of tasks.

To generate this I used the following code after loading a "rails console":

    puts Task.descendants.map(&:name).map(&:underscore).map{ |f| "require_dependency '#{f}'" }

Dear future reader of this commit, don't judge me too harshly.

APERTA-3399