Making sure Creators and Collaborators cannot access Production Metadata card.
This is the first permission for Creator that breaks the general rule that creators have view/edit access to all tasks. So now we can no longer rely on "Task" and we have to be more specific about which tasks they can access.

This also moves up the view/add/remove participants permission to be more specific.

APERTA-5635