Removing features/roles_and_permissions_spec
Tad and I talked about pushing the details of this down into unit tests so the test suite can avoid incurring the set up cost of all of these scenarios.

Since we decided to not go in the direction of using feature specs for testing the R&P matrix and given that this adds several minutes to each test run I'm nixing it now. 

We can revert this commit is we decide this was not a good decision.