Adding :pristine_roles_and_permissions flag for specs
Use :pristine_roles_and_permissions when your test should have a blank
slate for R&P. When running feature specs and unit specs in random order
feature specs may have pre-loaded R&P as a speed optimization and this may
cause some R&P things to already exist when the unit test runs.

Ideally, feature specs are updated to be optimized in a way that doesn't interfere with unit tests that come after. Will test an idea for that in a separate branch.

APERTA-5648