The RegisterDecisionTask no longer tries to pre-fill the AE's name
  * push template application out to the front-end where we have easy access to the current user

This is because there can now be more than one AE assigned to a paper and we don't know which one is responsible for sending the letter by default.

After talking with Glenn the RegisterDecision letter(s) should default to the currently logged in user's name (with no changes to the person's title in the closing).

APERTA-6252