Make sure selected user is cleared when changing remote data source.
For example, on the Assign Team card if you select a role and then a user... if you change the role to a different role you should be forced to select a new user.

* Clear selectedUser in the assign-team-task component
* Remove unused 'users' property reference in the assign-team-task component
* update select-2-single to force the user selection to be cleared the the remote data source changed

Note: the change to select-2-single was required otherwise the run/event loop seemed to lag one run loop cycle behind in ember... meaning that it would take you to change the selected role twice before the user was actually cleared.

APERTA-6344