WIP: Delegate states
This allows our roles & permissions system to delegate the permissible
states to a related object. For example, a task level permission might
allow a user to edit a task in a selected set of states, but those
states are part of the paper, not the task. This allows the task to
"delegate" to paper its state.