Adding explicit require_dependency in app/**/*.rb files when they make reference to constants defined in lib/
This is to avoid lack of thread-safety with Rails auto-loading as described in https://github.com/rails/rails/issues/13142

This issue has been resolved in Rails as of this commit: https://github.com/rails/rails/commit/a71350cae0082193ad8c66d65ab62e8bb0b7853b

But that commit is not available until Rails v5.0.0.beta1 and later at the time of this commit.

NOTES FROM INITIAL INVESTIGATION:

This solution works because the app is eager_loaded and everything inside of the app/ directory is loaded when the application boots. This loads EventStream::Notifiable and in turn will load Notifier.

To see the issue in action you need to load "rails console" before this commit:

 * Edit lib/notifier.rb and put "sleep 5" right inside the "class Notifier" declaration
 * Open rails console
 * In console verify that Notifier is NOT loaded: Object.constants.map(&:to_s).grep /Notifier/
 * In console re-create issue:

      arr = []
      arr << Thread.new { puts Notifier }
      arr << Thread.new { puts Notifier }
      arr.map(&:join)

This is not related to the Notifier name. The constant could be called anything else and this issue would still fail.

It seems that Ruby (or Rails, but most likely Ruby) is internally keeping track of what constant is currently being auto-loaded. When the second thread attempts to auto-load the same constant (while the first thread hasn't finished loading it yet) the "Circular dependency" error is raised.

Paired with Sam Bleckley (@diiq) on figuring this out.

APERTA-6537