Add SnapshotsById
When diffing figures, Authors, and Supporting information, among others, we face a funny problem when an item is deleted. If we just compare each pair of snapshots in order, the items get mis-aligned, and every item after the deleted one shows a diff, even when the item is actually unchanged. So we must pair our snapshots up by *id*, not by position. This is a helper class to make that pairing fast and easy.

See supporting-information-file-snapshot/component.js to see it in action.

Paired with Sam Bleckley <sam@mutuallyhuman.com>