Move search permissions to Paper
EligibleUserService.eligible_users won't filter out users that
have already been assigned to the task, which is bad. We need to change
either the logic or the assignment, and since requirement for the
permission is ambiguous from a business standpoint it's simpler to
assign the search permissions to the Paper.