Merge permission states instead of overwriting permission with the last one iterated over.
This ensures that when a user has a permission to an object thru multiple permission/role assignments that we merge the permission states rather than overwrite them.

If a user creates a paper and is also the Staff Admin on the journal they created the paper in then they will have the most generous permissions on their paper. 

Prior to this change that user may have add the Staff Admin permission states or may have had the Creator permission states. It was totally order dependent on the assignments coming back from the DB and being added to the ResultSet.

This makes the behavior predictable and consistent even if it is to change down the road.

APERTA-6137