Nest manuscript upload route under task
Permissions for uploading a manuscript depend on the ability to edit
the UploadManuscriptTask, and by nesting under the associated task we
can take advantage of the existing permission structure.  The other use
case for uploading a document - creating a new manuscript - is handled
by saving the paper with a new s3 url.