Resolving failing  spec/features/manuscript_new_spec.rb
In order to go full circle ihat needed to post back to Aperta the paper was done processing. So his commit includes a FakeIhatService for making real HTTP requests with dummy data. 

This required a new VCR request in manuscrupt-new.yml to represent downloading an ePub version of the paper from ihat. 

To generate that request that I ran the following code in a console session while in the test environment:

    VCR.use_cassette('dummy', :record => :new_episodes) do
      RestClient.get "https://s3-us-west-2.amazonaws.com/pressbooks-samplefiles/MetamorphosisJacksonTheme/Metamorphosis-jackson.epub"
    end

The above saved dummy.yml with the recorded to "spec/fixtures/vcr_cassettes/dummy.yml". I copied the request out of that file into manuscript-new.yml, then changed the URL.

APERTA-6597